+++ __tests__/outdatedPackages.test.ts
@@ -1,4 +1,4 @@
-import * as outdated from '../src/outdated'
+import * as outdated from '../src/outdatedPackages'
 import * as interfaces from '../src/interfaces'
 
 describe('splitAndRemoveEmptyString', () => {
